+++ packages/recommend-js/src/relatedProducts.tsx
@@ -7,7 +7,6 @@ import {
 import {
   createRelatedProductsComponent,
   RelatedProductsProps as RelatedProductsVDOMProps,
-  VNode,
 } from '@algolia/recommend-vdom';
 import { createElement, Fragment, h, render } from 'preact';
 import { useEffect, useState } from 'preact/hooks';
@@ -73,10 +72,7 @@ export function relatedProducts<TObject>({
   headerComponent,
   view,
   ...props
-}: RelatedProductsProps<TObject, Template> & EnvironmentProps):
-  | VNode
-  | VNode[]
-  | undefined {
+}: RelatedProductsProps<TObject, Template> & EnvironmentProps) {
   const children = (
     <RelatedProducts<TObject, Template>
       {...props}