+++ packages/recommend-js/src/trendingFacets.tsx
@@ -7,7 +7,6 @@ import {
 import {
   createTrendingFacetsComponent,
   TrendingComponentProps as TrendingFacetsVDOMProps,
-  VNode,
 } from '@algolia/recommend-vdom';
 import { createElement, Fragment, h, render } from 'preact';
 import { useEffect, useState } from 'preact/hooks';
@@ -72,10 +71,7 @@ export function trendingFacets<TObject>({
   headerComponent,
   view,
   ...props
-}: TrendingFacetsProps<TObject, Template> & EnvironmentProps):
-  | VNode
-  | VNode[]
-  | undefined {
+}: TrendingFacetsProps<TObject, Template> & EnvironmentProps) {
   const children = (
     <TrendingFacets<TObject, Template>
       {...props}