+++ packages/recommend-vdom/src/DefaultHeader.tsx
@@ -5,6 +5,10 @@ import { cx } from './utils';
 
 export function createDefaultHeaderComponent({ createElement }: Renderer) {
   return function DefaultHeader<TObject>(props: ComponentProps<TObject>) {
+    if (!props.recommendations || props.recommendations.length < 1) {
+      return null;
+    }
+
     if (!props.translations.title) {
       return null;
     }