+++ packages/recommend-vdom/src/FrequentlyBoughtTogether.tsx
@@ -58,6 +58,8 @@ export function createFrequentlyBoughtTogetherComponent({
           itemComponent: props.itemComponent,
           items: props.items,
           translations,
+          Fragment,
+          createElement,
           // @ts-ignore
           ...viewProps,
         })}