+++ packages/recommend-vdom/src/TrendingFacets.tsx
@@ -54,6 +54,8 @@ export function createTrendingFacetsComponent({
           itemComponent: props.itemComponent,
           items: props.items,
           translations,
+          Fragment,
+          createElement,
           // @ts-ignore
           ...viewProps,
         })}